design issues

+++ webapp/backend/src/server/services/items.service.js
@@ -89,7 +89,7 @@ class UsersService {
                 left join order_items oi on io.id = oi.item_option_id
                 left join orders o on o.id = oi.order_id
                 where (title like $1 or $1 is NULL) 
-                and ($6 is not NULL)
+                and ($6::int is not NULL)
                 and (o.status = 'basket' or o.status is NULL)  
                 and (p.category_name = $2 or $2 is NULL)  
                 and (p.id = $3 or $3 is NULL)
